成本库存-相关操作

+++ EMT.DoneNOW.Core/IVT/ivt_warehouse.cs
@@ -23,7 +23,7 @@ namespace EMT.DoneNOW.Core
         [DataMember]
         public SByte is_active { get; set; }
         [DataMember]
-        public Int32? resource_id { get; set; }
+        public Int64? resource_id { get; set; }
 
 
     }
